Moving around Module and VerifyKeyWrapper + some quality of life refactoring
